+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 - July 2022
   - Recycle 1.4.0
-    - Added `Restore-Item`
+    - Added `Restore-Item` and `Get-RecycledItem`
   - AdministratorRole 1.1.0
     - Added `Invoke-Elevated`
 - April 2022